+++ src/Commands/Enums/TeamsTemplateType.cs
@@ -1,0 +1,10 @@
+namespace PnP.PowerShell.Commands.Enums
+{
+
+    public enum TeamsTemplateType
+    {
+        None,
+        EDU_Class,
+        EDU_PLC,
+    }
+}